general: refactoring
